estimator/page refactoring
