+++ docs/authors.rst
@@ -5,7 +5,7 @@ Credits
 Development
 -----------
 
-* Leszek Grzanka - IFJ-PAN, Poland <leszek.grzanka@gmail.com>
+* Leszek Grzanka - IFJ-PAN, Poland <leszek.grzanka@ifj.edu.pl>
 * Niels Bassler - Stockholm University, Sweden
 
 Contributors
