+++ pymchelper/readers/shieldhit/reader_bdo2016.py
@@ -2,7 +2,8 @@ import logging
 
 import numpy as np
 
-from pymchelper.estimator import MeshAxis, Page
+from pymchelper.axis import MeshAxis
+from pymchelper.page import Page
 from pymchelper.readers.shieldhit.reader_base import SHReader, _get_mesh_units, _bintyp, _get_detector_unit, \
     read_next_token
 from pymchelper.readers.shieldhit.binary_spec import SHBDOTagID, detector_name_from_bdotag
