+++ pymchelper/readers/shieldhit/reader_bin2010.py
@@ -3,7 +3,8 @@ import logging
 
 import numpy as np
 
-from pymchelper.estimator import MeshAxis, Page
+from pymchelper.axis import MeshAxis
+from pymchelper.page import Page
 from pymchelper.readers.shieldhit.reader_base import SHReader, _get_mesh_units, _bintyp, _get_detector_unit
 from pymchelper.shieldhit.detector.detector_type import SHDetType
 from pymchelper.shieldhit.detector.estimator_type import SHGeoType
